refactor (feat): loop and cue safety (#14)
PR replaces the CUE button with a LOOP button which sets a loop between
the two hot cue points.
PR also implements cue safety so that a cue point is not triggered to
play a deck when the other deck is playing. The check is the following:

- other deck is not playing -> start cue
- other deck has volume zero (prelistening) -> start cue
- this deck has volume zero - > start cue
- none of the above ? -> Do nothing 

The hot cues will light up when they can be activated.