Update single-dual-deck switch (#19)
PR updates behavior when switching from single to dual deck using
button.

Issue was that in previous behavior, if the second deck had a song that
was not part of the automix, then when switching from single to dual
deck this song would be cued next.

PR updates behavior to load next automix song on the idle deck when
switching to dual deck