The video preview has been enlarged.
The video preview which appears when you left click the video button
has been made more than twice as wide and twice as high. It now covers
the master section. Right-clicking the preview will hide it, while right-
clicking the video button will cause it to reappear. The actual video
display remains open and active while the preview is hidden.

Previously, right-clicking on the video button or the video preview
would open the video-output selector. This has been assigned to a
double-click on the video button or video preview, though it looks like
double-clicking (the dblclick attribute) doesn't work properly in
VirtualDJ since the first click still fires and causes video to be
toggled between open and closed.