Incorporate-large-preview-toggle (#25)
PR implements great proposal suggested in #23 of allowing video preview
to be larger along with fixing a bug introduced in pull #24.

In #23 Hartag notes that "The video preview in TigerTango is too small
to be able to clearly read information displayed on the external
screen." Solution is to allow a larger preview to be toggled.

This PR implements a slightly different logic to get the same effect. 
By switching the video output selection button with the video effects
button (to allow things like text to be added to video), the only way to
get the video output is the right click on the video area. To make this
easier to do, this PR implements the following logic:
* If no video is active, then right click opens the video output
selection menu.
* If video is active, then right click toggles the larger video screen. 
* If you click on the video then it will turn off the video preview,
leaving the size toggle at whatever was previously set (so if you set to
larger preview and then close preview, next time you open it will be
larger).