Update auto button default (#33)
PR updates auto button such that if no song is loaded onto deck then it
defaults to the first automix song.

Previous button was defaulting to load whatever was highlighted in
browser at the first automix if the deck had nothing loaded. This could
cause confusion, hence the update.

This means AUTO now uses the following logic:

- Other deck playing? have that song as first automix and load next song
in list
- No track loaded --> load first song in automix list
- Track loaded ? load this to automix and have it be first in line.