feat: Video skin updates (#39)
This PR makes several updates to the interface between TigerTango and
TigerTangoVideo to enable displaying of information to the video skin.
The new functionality is as follows: 

1. based on all of the work from @hartag , this updated skin includes a
counter for the position in the tanda. This counter uses the Genre tag
to determine which type of music is playing, and takes settings such as
the number of songs in a Tango or Vals tanda. The documentation has been
updated to describe this new capability.
2. Song lyrics can now be displayed. Lyrics are taken from the Comments
metadata tag (ideally could have a unique 'lyrics' tag, but it seems
currently comments is the only metadata tag that allows multiple lines.
3. A range of settings are enabled to customize the video look
4. The video skin has been updated in appearance. 
5. Documentation has been updated