feat: Deactivate dropzone when in automix (#55)
PR adds a safety measure to deactivate the drop zone areas when in
automix.

The drop zones allow you to drag and drop tracks directly to the
turntables. This is useful when automix is off, but can easily cause
issues when automix is on. The ideal is for users to add tracks to the
automix list when automix is active.

The PR removes the dropzones when automix is active, disabling adding
tracks directly to the turntables, and adds a small button icon to
signify as much.

PR also cleans up code base by removing jogwheel option