feat, refactor: update skin layout (#57)
PR makes a number of changes to the skin layout and adds additional
functionality.

Main update is to move pieces that affect the full deck, namely the
toolbox and windows areas, up to the top area with the rest of the
master effects. This allows the two deck areas to be symmetric and opens
up space.

Updates include:

* Update frequency sliders to knobs to make layout look cleaner and
symmetric.
* Making warning lights only appear then there is an issue 
* Add save and color buttons
* Expose automix options
* Move windows to a menu
* Update waveform window